check for existence of video path else assume relative path

+++ dandi/organize.py
@@ -259,6 +259,11 @@ def organize_external_files(
                     continue
                 new_path = op.join(dandiset_path, op.dirname(e["dandi_path"]), name_new)
                 name_old_str = str(name_old)
+                if not op.isabs(name_old_str):
+                    name_old_str = op.join(op.dirname(e["path"]), name_old_str)
+                if not op.exists(name_old_str):
+                    lgr.error("%s does not exist", name_old_str)
+                    raise FileNotFoundError(f"{name_old_str} does not exist")
                 os.makedirs(op.dirname(new_path), exist_ok=True)
                 if files_mode == "symlink":
                     os.symlink(name_old_str, new_path)
